[ci]: Keep standalone validation additive
