+++ azure-pipelines.yml
@@ -111,6 +111,10 @@ stages:
       clean: true
       displayName: 'Checkout code'
 
+    - checkout: sonic-mgmt-common
+      clean: true
+      displayName: 'Checkout sonic-mgmt-common'
+
     - template: .azure/templates/install-go.yml
       parameters:
         version: $(GO_VERSION)
@@ -118,6 +122,8 @@ stages:
     - bash: |
         set -euo pipefail
         export PATH=$PATH:/usr/local/go/bin:$(go env GOPATH)/bin
+        cd sonic-gnmi
+        go mod tidy
         make -f pure.mk junit-xml
       displayName: 'Run pure package tests'
 
@@ -126,7 +132,7 @@ stages:
       condition: always()
       inputs:
         testResultsFormat: 'JUnit'
-        testResultsFiles: '$(Build.SourcesDirectory)/test-results/junit-pure.xml'
+        testResultsFiles: '$(System.DefaultWorkingDirectory)/sonic-gnmi/test-results/junit-pure.xml'
         failTaskOnFailedTests: true
         publishRunAttachments: true
         testRunTitle: 'Pure package tests'
@@ -135,9 +141,9 @@ stages:
       displayName: 'Publish pure package coverage'
       condition: always()
       inputs:
-        summaryFileLocation: '$(Build.SourcesDirectory)/test-results/coverage-pure.xml'
+        summaryFileLocation: '$(System.DefaultWorkingDirectory)/sonic-gnmi/test-results/coverage-pure.xml'
 
-    - publish: $(Build.SourcesDirectory)/test-results/coverage-pure.xml
+    - publish: $(System.DefaultWorkingDirectory)/sonic-gnmi/test-results/coverage-pure.xml
       artifact: coverage-pure
       displayName: 'Publish pure coverage artifact'
       condition: always()
@@ -207,7 +213,6 @@ stages:
       inputs:
         testResultsFormat: 'JUnit'
         testResultsFiles: |
-          $(System.DefaultWorkingDirectory)/sonic-gnmi/test-results/junit-integration-dbconfig.xml
           $(System.DefaultWorkingDirectory)/sonic-gnmi/test-results/junit-integration-basic.xml
           $(System.DefaultWorkingDirectory)/sonic-gnmi/test-results/junit-integration-env.xml
           $(System.DefaultWorkingDirectory)/sonic-gnmi/test-results/junit-integration-dialout.xml
